fix recale_position calls, MURDUR declaration

+++ include/collision/block_6E5E0.h
@@ -11,7 +11,7 @@
 #include "ray/ray_5D190.h"
 #include "command_56AF0.h"
 
-u8 MURDUR(s16 x, s16 y);
+s32 MURDUR(s16 x, s16 y); /* s16 or RayTestBlocSH needs casts */
 s16 dist_to_bloc_floor(s16 btyp, s16 x, s16 y);
 s16 bloc_floor(s16 btyp, s16 x, s16 y);
 s32 calc_typ_trav(Obj *obj, u8 param_2);
